+++ test/chemistry/test_swaprz.py
@@ -12,10 +12,13 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-""" Test of SWAPRZ from core aqua """
+"""Test of SWAPRZ from the circuit library."""
 
+import warnings
 from test.chemistry import QiskitChemistryTestCase
+from ddt import ddt, data
 from qiskit import BasicAer
+from qiskit.circuit.library import ExcitationPreserving
 from qiskit.aqua import QuantumInstance, aqua_globals
 from qiskit.aqua.algorithms import VQE
 from qiskit.aqua.components.optimizers import SLSQP
@@ -25,6 +28,7 @@ from qiskit.chemistry.drivers import HDF5Driver
 from qiskit.chemistry.core import Hamiltonian, QubitMappingType
 
 
+@ddt
 class TestSwapRZ(QiskitChemistryTestCase):
     """
        SwapRZ was designed to preserve particles. We test it here from
@@ -39,7 +43,8 @@ class TestSwapRZ(QiskitChemistryTestCase):
         aqua_globals.random_seed = self.seed
         self.reference_energy = -1.137305593252385
 
-    def test_swaprz(self):
+    @data('wrapped', 'library')
+    def test_swaprz(self, mode):
         """ SwapRZ variational form test """
 
         driver = HDF5Driver(self.get_resource_path('test_driver_hdf5.hdf5'))
@@ -54,8 +59,18 @@ class TestSwapRZ(QiskitChemistryTestCase):
                                     operator.molecule_info['num_particles'],
                                     qubit_mapping=operator._qubit_mapping,
                                     two_qubit_reduction=operator._two_qubit_reduction)
-        var_form = SwapRZ(qubit_op.num_qubits, initial_state=initial_state)
-        algo = VQE(qubit_op, var_form, optimizer)
+
+        if mode == 'wrapped':
+            warnings.filterwarnings('ignore', category=DeprecationWarning)
+            wavefunction = SwapRZ(qubit_op.num_qubits, initial_state=initial_state)
+        else:
+            wavefunction = ExcitationPreserving(qubit_op.num_qubits, initial_state=initial_state)
+
+        algo = VQE(qubit_op, wavefunction, optimizer)
+
+        if mode == 'wrapped':
+            warnings.filterwarnings('always', category=DeprecationWarning)
+
         result = algo.run(QuantumInstance(BasicAer.get_backend('statevector_simulator'),
                                           seed_simulator=aqua_globals.random_seed,
                                           seed_transpiler=aqua_globals.random_seed))